Tag edge deletion
With this, the hierarchical tagging system should be complete.
We'll want to use a library for displaying trees to make a
component for displaying, modifying, organizing tags etc.
That will involve new queries but they should be simple, as
inserting and deleting edges are the only complex queries
for this model.

We can then easily hook up the tags view to filter the table of
results so we only view photos with certain tags.