Working inference from model
Shows a working inference that matches the expectation from the
OpenAI CLIP algorithm. From a small set of informat tests, it
is giving the confidence rankings for sets of text appropriately.

This shows many text queries and their likelihood for a single
image. We will want to invert that and use one NL query
across many images, and rank the images. But the legwork is
the same.

We'll move with a naive approach of something like "run inference
for every image" at first for the MVP and see how it performs.
I think we will likely need some caching strategy, but folks
have made stuff like this before so I am sure there are
sufficient strategies available.