Add Clip.encode_text()
We're simplifying to only use ONNX and not doing any tensor ops
in this app. Doing so would require a tch-rs dependency or
a bespoke solution, and just shipping the full ONNX graphs is
actually smaller, even with duplicate logic.

Plus I figured out how to make the encode_text() export work by
just wrapping it in a nn.Module within the exporter, so that made
things much easier.