Add basic display of images on search.
We will want to move towards a thumbnail-based approach and a proper
gallery, but this is a start.

Displaying arbitrary images on the filesystem on the JS side is
actually non-trivial in Tauri. It's... annoying.
We instead return a base64 encoding from Rust.
This will change, and we'll do the base64 encoding technique only
if we need the full resolution for some reason (which should be rare).