Place Search into PanelGroup layout
This simplifies the UI a bit, since it removes the outer scroll
easily. But, it means the search bar div takes up 9% of the
window always, which means it grows/shrinks if the window is
resized. For that reason, I might tinker further as I'd want
the search bar to keep a certain pixel height, especially to
ensure the search bar itself isn't clipped, as it currently does
if the window is very short.