Handle and track failed encodings
Mostly, this just means "it's not an image or it's formatted wrong".
We track these so we can choose to handle them later; if we make
a change that adds a new filetype support, for example, we could
scan through and try encoding them again. Or something - it's not
a pressing issue right now.