Add TaskError and handle it on frontend
Before (and still in some cases to be fixed in an upcoming change)
we wouldn't handle errors within commands more than logging them.
This meant that we did not receive TaskEnd events to match
TaskStatus events given prior, so these would not be cleared
out of the TaskStatuses map.

This change returns the task ID from errors within the
add_watched_directory command, so that the frontend can (and
does, in this change) remove references to that task,
effectively acting as a "TaskEnd" event.

In particular, this makes the list of active taskstatuses
from the rover-analyzer button not just keep displaying
failed tasks.